update librespot 0.8.0
Fix: Delete the connect state only on spirc shutdown #1555

In #1551 @samcook encountered an unexpected behavior during reconnecting.

The issue happens because the reconnect is initiated via zeroconf instead of using the connect interface, aka in the disconnect handling was a major issue where the state itself was deleted. This should only happen when shutting down the dealer so that the dealer can still receive requests and updates.